feat: add user repository (#21)
* feat: add user repository
* refactor(user.repository.ts): move s3 dependency into user service
---------
