+++ src/prisma/types.ts
@@ -1,0 +1,8 @@
+export type FindResponse<T> = {
+  data: T[];
+  meta: {
+    skip: number;
+    take: number;
+    count: number;
+  };
+};
