+++ src/users/users.module.ts
@@ -4,10 +4,11 @@ import { UsersController } from './users.controller';
 import { PrismaModule } from '../prisma/prisma.module';
 import { BcryptService } from '../service/bcrypt.service';
 import { S3Service } from '../service/s3.service';
+import { UsersRepository } from './repository/users.repository';
 
 @Module({
   controllers: [UsersController],
-  providers: [UsersService, BcryptService, S3Service],
+  providers: [UsersService, BcryptService, S3Service, UsersRepository],
   imports: [PrismaModule],
   exports: [UsersService],
 })
