+++ src/users/users.service.ts
@@ -1,28 +1,25 @@
-import { extname } from 'path';
-import { createHash } from 'crypto';
-import { Role } from '@prisma/client';
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { User } from '@prisma/client';
+import { Injectable } from '@nestjs/common';
 
 import { FindUserDto } from './dto/find-user.dto';
-import { S3Service } from '../service/s3.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { PrismaService } from '../prisma/prisma.service';
 import { BcryptService } from '../service/bcrypt.service';
 import { ICreateStrategy } from './create-strategy/icreate.strategy';
 import { UploadProfilePhotoResponse } from './dto/upload-avatar-response.dto';
+import { UsersRepository } from './repository/users.repository';
+import { S3Service } from '../service/s3.service';
+import { extname } from 'path';
+import { createHash } from 'crypto';
 
-type UserAttributesFilter = {
-  userAttributes: {
-    [key: string]: { startsWith: string; mode: 'insensitive' };
-  };
-};
 @Injectable()
 export class UsersService {
   constructor(
-    private prismaService: PrismaService,
-    private s3Service: S3Service,
-    public bcryptService: BcryptService,
+    private readonly prismaService: PrismaService,
+    private readonly s3Service: S3Service,
+    private readonly usersRepository: UsersRepository,
+    private readonly bcryptService: BcryptService,
   ) {}
 
   async create(createUserDto: CreateUserDto, strategy: ICreateStrategy) {
@@ -33,66 +30,15 @@ export class UsersService {
     return strategy.create({ ...createUserDto, password });
   }
 
-  async find(query: FindUserDto) {
-    const take = query.limit ? Number(query.limit) : 10;
-    const skip = query.page ? Number(query.page) * take : 0;
-    const orderBy =
-      query.sort && (query.dir === 'asc' || query.dir === 'desc')
-        ? { [query.sort]: query.dir }
-        : null;
-    const filter = query.filter
-      ? Object.entries(query.filter).reduce<UserAttributesFilter>(
-          (filterObject: UserAttributesFilter, [key, value]) => {
-            filterObject.userAttributes[key] = {
-              startsWith: value,
-              mode: 'insensitive',
-            };
-            return filterObject;
-          },
-          { userAttributes: {} },
-        )
-      : null;
-    const [count, data] = await this.prismaService.$transaction([
-      this.prismaService.user.count({
-        where: {
-          role: 'USER',
-          active: true,
-          ...(filter && { ...filter }),
-        },
-      }),
-      this.prismaService.user.findMany({
-        skip,
-        take,
-        ...(orderBy && { orderBy }),
-        where: { role: 'USER', active: true, ...(filter && { ...filter }) },
-        include: { userAttributes: true },
-      }),
-    ]);
-
-    return { data, meta: { skip, take, count } };
+  async find(query: FindUserDto, user: User) {
+    return this.usersRepository.find(query, user);
   }
 
   async findOne(id: number) {
-    const user = await this.prismaService.user.findUnique({
-      where: { id },
-      select: {
-        userAttributes: true,
-        organisationAttributes: true,
-        id: true,
-        role: true,
-        email: true,
-        active: true,
-        profilePhotoKey: true,
-      },
-    });
-
-    if (!user) {
-      throw new NotFoundException();
-    }
-
+    const user = await this.usersRepository.findOne(id);
     const { profilePhotoKey, ...restUser } = user;
 
-    let profilePhoto = null;
+    let profilePhoto = '';
     if (profilePhotoKey) {
       profilePhoto = await this.s3Service.get(profilePhotoKey);
     }
@@ -108,95 +54,33 @@ export class UsersService {
   }
 
   async update(id: number, updateUserDto: UpdateUserDto) {
-    const user = await this.prismaService.user.findUnique({
-      where: { id },
-      select: { role: true },
-    });
-
-    if (!user) throw new NotFoundException();
-
-    const userData = { active: updateUserDto.active };
-
-    const userAttributesData = {
-      firstname: updateUserDto.firstname,
-      lastname: updateUserDto.lastname,
-      city: updateUserDto.city,
-      phone: updateUserDto.phone,
-      type: updateUserDto.type,
-    };
-
-    const organisationAttributesData = {
-      name: updateUserDto.name,
-      street: updateUserDto.street,
-      oib: updateUserDto.oib,
-      city: updateUserDto.city,
-    };
-
-    return this.prismaService.user.update({
-      where: { id },
-      data: {
-        ...userData,
-        ...(user.role === Role.USER && {
-          userAttributes: { update: { ...userAttributesData } },
-        }),
-        ...(user.role === Role.ORGANISATION && {
-          organisationAttributes: { update: { ...organisationAttributesData } },
-        }),
-      },
-      select: {
-        userAttributes: true,
-        organisationAttributes: true,
-        id: true,
-        email: true,
-        active: true,
-      },
-    });
+    return this.usersRepository.update(id, updateUserDto);
   }
 
   async remove(id: number) {
-    const user = await this.prismaService.user.findUnique({
-      where: { id },
-      select: {
-        userAttributes: { select: { certificates: { select: { key: true } } } },
-        email: true,
-      },
-    });
+    const user = await this.usersRepository.remove(id);
+
     if (user?.userAttributes?.certificates?.length) {
       const keys = user.userAttributes.certificates.map(({ key }) => key);
       await this.s3Service.deleteMany(keys);
     }
-
-    return this.prismaService.user.delete({ where: { id } });
   }
 
   async uploadProfilePhoto(
     id: number,
     file: Express.Multer.File,
   ): Promise<UploadProfilePhotoResponse> {
-    const user = await this.prismaService.user.findUniqueOrThrow({
-      where: { id },
-      select: {
-        id: true,
-      },
-    });
-
     const fileExt = extname(file.originalname);
-    const filename = createHash('md5').update(String(user.id)).digest('hex');
+    const filename = createHash('md5').update(String(id)).digest('hex');
     const profilePhotoKey = `profile_photos/${filename}${fileExt}`;
 
     await this.s3Service.upload(file.buffer, profilePhotoKey, {
       mimetype: file.mimetype,
     });
 
-    await this.prismaService.user.update({
-      where: { id },
-      data: {
-        profilePhotoKey,
-      },
-    });
+    await this.usersRepository.updateProfilePhoto(id, profilePhotoKey);
 
     const profilePhoto = await this.s3Service.get(profilePhotoKey);
-
     return {
       profilePhoto,
     };
